Moved the cache_purge test into linux
